+++ docs/internal/processes/development-guide-relevant-files.md
@@ -163,6 +163,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/config/layouttests/layout_clone_test.go` | Layout clone behavioral regression surface | `CloneWorkerConfig(...)` and `CloneFactoryConfig(...)` regression coverage should assert observable detached-copy behavior from an external test package here by mutating the source config after cloning and checking the cloned outcome, rather than asserting helper placement, helper names, or source-file inventories. |
 | `pkg/config/maptests/`, `pkg/config/mappingtests/`, `pkg/config/runtimetests/`, and `pkg/config/exhaustiontests/` | Config package-size test split seam | Keep exported-surface mapper, generated-boundary mapping, runtime-load, named-factory, and contract-guard coverage in focused external test packages when root `pkg/config` approaches the file-count cap. Keep root config tests only for package-private semantic validator, layout materialization, and staged persistence helpers that should not be exported for test movement. |
 | `pkg/service/modeltests/` | Service model API package-size test split seam | Keep black-box coverage for exported model read APIs such as `FactoryService.ListModels(...)` and `GetModel(...)` in this external test package, building the service through disk-backed factory config rather than assigning package-private runtime fields. Keep root `pkg/service` model tests only when they need package-private puller, local-runtime, resource-limiter, or invocation executor seams. |
+| `pkg/service/*_test.go` | Service same-package package-size cleanup seam | Consolidate only small same-package service tests into same-concern owners while each file remains below the backend 1000-line size gate. Tests that assert package-private runtime handles, service fields, poller internals, local-model manager internals, or helper functions should stay same-package until the owning production seam is split; do not move them to external packages just to satisfy package-size lint. |
 | `pkg/config/mappingtests/` | Generated factory-config boundary regression surface | Boundary rejection regressions should stay focused here by asserting the generated-boundary error prefix plus the full path-specific retirement guidance string for top-level, nested `definition`, and nested `definition.cron` aliases. |
 | `pkg/config/openapitests/` | Generated OpenAPI JSON boundary regression surface | When cleanup work must preserve generated-boundary error strings or prove the generated `Factory` JSON shape itself, keep black-box proof here with `GeneratedFactoryFromOpenAPIJSON(...)` assertions over the generated-boundary error prefix, exact field paths, replacement guidance, and marshaled `factoryapi.Factory` workstation keys. |
 | `pkg/replay/event_stream_artifact.go` | Replay event-stream import and adjacent-factory reconstruction owner | SSE replay import, truncated-tail recovery, legacy cron compatibility normalization, and adjacent authored-factory hydration plus payload rewriting should stay centralized here so saved event streams rebuild the same backend-visible factory shape as canonical replay artifacts; tests that need file-based conversion should call `ArtifactFromEventStreamFile` or `SaveArtifactFromEventStreamFile` instead of copying that hydration path into suite-local helpers. |
