+++ pkg/service/runtime_session_runtime_test.go
@@ -4,11 +4,13 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 	"time"
 
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
 	"github.com/portpowered/infinite-you/pkg/apisurface"
 	"github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/factory/requests"
@@ -429,5 +431,327 @@ func startRunningSessionServiceOnDir(t *testing.T, rootDir string) *runningSessi
 		svc:       svc,
 		runErrCh:  runErrCh,
 		cancelRun: cancelRun,
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_AutoOpensSingleTarget(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		defaultFactory: "alpha",
+		namedFactories: []string{"alpha"},
+	})
+	defer harness.stop(t)
+
+	result, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, nil, false)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(single target): %v", err)
+	}
+	if result == nil || result.SessionID == "" {
+		t.Fatalf("single-target open result = %#v, want session id", result)
+	}
+	if len(result.Targets) != 0 {
+		t.Fatalf("single-target open returned picker targets = %#v, want none", result.Targets)
+	}
+	session := harness.requireSession(t, result.SessionID)
+	if session.handle.runtime.dir != harness.factoryDirs["alpha"] {
+		t.Fatalf("opened session runtime dir = %q, want %q", session.handle.runtime.dir, harness.factoryDirs["alpha"])
+	}
+	if got := harness.svc.sessions.count(); got != 2 {
+		t.Fatalf("live session count = %d, want 2", got)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_ReturnsTargetPickerMetadata(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		rootConfig:     minimalFactoryConfig(),
+		namedFactories: []string{"alpha", "beta"},
+	})
+	defer harness.stop(t)
+
+	result, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, nil, false)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(multi target): %v", err)
+	}
+	if result == nil {
+		t.Fatal("expected target picker result, got nil")
+	}
+	if result.SessionID != "" {
+		t.Fatalf("multi-target open returned session %q, want target picker", result.SessionID)
+	}
+	if len(result.Targets) != 3 {
+		t.Fatalf("target picker count = %d, want 3", len(result.Targets))
+	}
+
+	assertSessionTargetMetadata(t, result.Targets[0], FactorySessionTargetKindDefault, "", "default", harness.rootDir, "factory")
+	assertSessionTargetMetadata(t, result.Targets[1], FactorySessionTargetKindNamed, "alpha", "alpha", filepath.Join(harness.rootDir, "alpha"), "alpha")
+	assertSessionTargetMetadata(t, result.Targets[2], FactorySessionTargetKindNamed, "beta", "beta", filepath.Join(harness.rootDir, "beta"), "beta")
+	if got := harness.svc.sessions.count(); got != 1 {
+		t.Fatalf("target-picker flow mutated live sessions to %d, want 1", got)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_OpensExplicitDefaultAndNamedTargets(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		rootConfig:     minimalFactoryConfig(),
+		namedFactories: []string{"beta"},
+	})
+	defer harness.stop(t)
+
+	defaultOpen, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, &FactorySessionTargetRef{
+		Kind: FactorySessionTargetKindDefault,
+	}, false)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(default): %v", err)
+	}
+	betaOpenOne, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, &FactorySessionTargetRef{
+		Kind: FactorySessionTargetKindNamed,
+		Name: "beta",
+	}, false)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(beta one): %v", err)
+	}
+	betaOpenTwo, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, &FactorySessionTargetRef{
+		Kind: FactorySessionTargetKindNamed,
+		Name: "beta",
+	}, false)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(beta two): %v", err)
+	}
+	if betaOpenOne.SessionID == betaOpenTwo.SessionID {
+		t.Fatalf("duplicate beta session ids = %q", betaOpenOne.SessionID)
+	}
+
+	defaultSession := harness.requireSession(t, defaultOpen.SessionID)
+	betaSessionOne := harness.requireSession(t, betaOpenOne.SessionID)
+	betaSessionTwo := harness.requireSession(t, betaOpenTwo.SessionID)
+	if defaultSession.handle.runtime.dir != harness.rootDir {
+		t.Fatalf("default target runtime dir = %q, want %q", defaultSession.handle.runtime.dir, harness.rootDir)
+	}
+	if betaSessionOne.handle.runtime.dir != harness.factoryDirs["beta"] || betaSessionTwo.handle.runtime.dir != harness.factoryDirs["beta"] {
+		t.Fatalf("beta target runtime dirs = %q and %q, want %q", betaSessionOne.handle.runtime.dir, betaSessionTwo.handle.runtime.dir, harness.factoryDirs["beta"])
+	}
+	if got := harness.svc.sessions.count(); got != 4 {
+		t.Fatalf("live session count = %d, want 4", got)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_RejectsInvalidFolderAndTargetWithoutMutation(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		rootConfig: minimalFactoryConfig(),
+	})
+	defer harness.stop(t)
+
+	before := harness.svc.sessions.count()
+	if _, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), filepath.Join(harness.rootDir, "missing"), nil, false); err == nil || !strings.Contains(err.Error(), "stat factory session folder") {
+		t.Fatalf("OpenFactorySessionFromFolder(missing folder) error = %v, want folder stat failure", err)
+	} else {
+		assertFactorySessionValidationTarget(t, err, "missing", "folderPath")
+	}
+	if got := harness.svc.sessions.count(); got != before {
+		t.Fatalf("missing-folder open mutated live sessions to %d, want %d", got, before)
+	}
+
+	if _, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, &FactorySessionTargetRef{
+		Kind: FactorySessionTargetKindNamed,
+		Name: "missing",
+	}, false); err == nil || !strings.Contains(err.Error(), `factory session target "missing" was not found`) {
+		t.Fatalf("OpenFactorySessionFromFolder(missing target) error = %v, want missing-target failure", err)
+	} else {
+		assertFactorySessionValidationTarget(t, err, "target_not_found", "target.name")
+	}
+	if got := harness.svc.sessions.count(); got != before {
+		t.Fatalf("missing-target open mutated live sessions to %d, want %d", got, before)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_RejectsReadableFolderWithoutRunnableTargetsWithoutMutation(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		rootConfig: minimalFactoryConfig(),
+	})
+	defer harness.stop(t)
+
+	before := harness.svc.sessions.count()
+	emptyDir := filepath.Join(harness.rootDir, "empty")
+	if err := os.Mkdir(emptyDir, 0o755); err != nil {
+		t.Fatalf("Mkdir(empty): %v", err)
+	}
+
+	if _, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), emptyDir, nil, false); err == nil || !strings.Contains(err.Error(), `does not expose any runnable factory targets`) {
+		t.Fatalf("OpenFactorySessionFromFolder(empty runnable folder) error = %v, want no-runnable-targets failure", err)
+	} else {
+		assertFactorySessionValidationTarget(t, err, "not_runnable", "folderPath")
+	}
+	if got := harness.svc.sessions.count(); got != before {
+		t.Fatalf("empty-folder open mutated live sessions to %d, want %d", got, before)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_CanceledRequestDoesNotRegisterSession(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		defaultFactory: "beta",
+		namedFactories: []string{"beta"},
+	})
+	defer harness.stop(t)
+
+	beforeIDs := harness.svc.sessions.ids()
+	if len(beforeIDs) != 1 || beforeIDs[0] != defaultFactorySessionID {
+		t.Fatalf("session ids before canceled open = %v, want [%s]", beforeIDs, defaultFactorySessionID)
+	}
+
+	openCtx, cancelOpen := context.WithCancel(context.Background())
+	cancelOpen()
+
+	if _, err := harness.svc.OpenFactorySessionFromFolder(openCtx, harness.factoryDirs["beta"], nil, false); !errors.Is(err, context.Canceled) {
+		t.Fatalf("OpenFactorySessionFromFolder(canceled) error = %v, want context canceled", err)
+	}
+
+	deadline := time.Now().Add(200 * time.Millisecond)
+	for time.Now().Before(deadline) {
+		if got := harness.svc.sessions.ids(); len(got) == 1 && got[0] == defaultFactorySessionID {
+			time.Sleep(10 * time.Millisecond)
+			continue
+		}
+		t.Fatalf("canceled open mutated live sessions to %v, want only [%s]", harness.svc.sessions.ids(), defaultFactorySessionID)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_ValidateOnlyReturnsTargetsWithoutOpening(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		defaultFactory: "alpha",
+		namedFactories: []string{"alpha"},
+	})
+	defer harness.stop(t)
+
+	before := harness.svc.sessions.count()
+	result, err := harness.svc.OpenFactorySessionFromFolder(context.Background(), harness.rootDir, nil, true)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(validate only): %v", err)
+	}
+	if result == nil {
+		t.Fatal("validate-only result = nil, want target metadata")
+	}
+	if result.SessionID != "" {
+		t.Fatalf("validate-only session id = %q, want none", result.SessionID)
+	}
+	if len(result.Targets) != 1 {
+		t.Fatalf("validate-only targets = %#v, want one target", result.Targets)
+	}
+	assertSessionTargetMetadata(t, result.Targets[0], FactorySessionTargetKindNamed, "alpha", "alpha", harness.factoryDirs["alpha"], "alpha")
+	if got := harness.svc.sessions.count(); got != before {
+		t.Fatalf("validate-only mutated live sessions to %d, want %d", got, before)
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_ExpandsLeadingTildeForValidationAndLaunch(t *testing.T) {
+	harness := startRunningSessionService(t, runningSessionServiceOptions{
+		defaultFactory: "alpha",
+		namedFactories: []string{"alpha"},
+	})
+	defer harness.stop(t)
+
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		t.Fatalf("UserHomeDir: %v", err)
+	}
+	relativeToHome, err := filepath.Rel(homeDir, harness.rootDir)
+	if err != nil {
+		t.Fatalf("filepath.Rel(home, root): %v", err)
+	}
+	if relativeToHome == "." || strings.HasPrefix(relativeToHome, "..") {
+		t.Skipf("root dir %q is not under the user home %q", harness.rootDir, homeDir)
+	}
+
+	tildePath := "~"
+	if relativeToHome != "." {
+		tildePath = filepath.Join("~", relativeToHome)
+	}
+
+	validateResult, err := harness.svc.OpenFactorySessionFromFolder(
+		context.Background(),
+		tildePath,
+		nil,
+		true,
+	)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(validate tilde): %v", err)
+	}
+	if validateResult == nil || len(validateResult.Targets) != 1 {
+		t.Fatalf("validate-only tilde targets = %#v, want one target", validateResult)
+	}
+	assertSessionTargetMetadata(
+		t,
+		validateResult.Targets[0],
+		FactorySessionTargetKindNamed,
+		"alpha",
+		"alpha",
+		harness.factoryDirs["alpha"],
+		"alpha",
+	)
+	if validateResult.Targets[0].FolderPath != harness.rootDir {
+		t.Fatalf("validated tilde folder path = %q, want %q", validateResult.Targets[0].FolderPath, harness.rootDir)
+	}
+
+	openResult, err := harness.svc.OpenFactorySessionFromFolder(
+		context.Background(),
+		tildePath,
+		&FactorySessionTargetRef{
+			Kind: FactorySessionTargetKindNamed,
+			Name: "alpha",
+		},
+		false,
+	)
+	if err != nil {
+		t.Fatalf("OpenFactorySessionFromFolder(open tilde): %v", err)
+	}
+	session := harness.requireSession(t, openResult.SessionID)
+	if session.folderPath != harness.rootDir {
+		t.Fatalf("opened session folder path = %q, want %q", session.folderPath, harness.rootDir)
+	}
+	if session.handle.runtime.dir != harness.factoryDirs["alpha"] {
+		t.Fatalf("opened session runtime dir = %q, want %q", session.handle.runtime.dir, harness.factoryDirs["alpha"])
+	}
+}
+
+func TestFactoryService_OpenFactorySessionFromFolder_InvalidExpandedTildePathReturnsResolvedError(t *testing.T) {
+	missingPath := filepath.Join("~", ".infinite-you-missing-factory-folder")
+
+	_, err := resolveFactorySessionFolder(missingPath)
+	if err == nil {
+		t.Fatal("resolveFactorySessionFolder(~missing) error = nil, want failure")
+	}
+
+	homeDir, homeErr := os.UserHomeDir()
+	if homeErr != nil {
+		t.Fatalf("UserHomeDir: %v", homeErr)
+	}
+	wantResolvedPath := filepath.Join(homeDir, ".infinite-you-missing-factory-folder")
+	if !strings.Contains(err.Error(), wantResolvedPath) {
+		t.Fatalf("resolveFactorySessionFolder(~missing) error = %q, want resolved path %q", err, wantResolvedPath)
+	}
+	assertFactorySessionValidationTarget(t, err, "missing", "folderPath")
+}
+
+func assertFactorySessionValidationTarget(t *testing.T, err error, wantReason string, wantField string) {
+	t.Helper()
+
+	var targetedErr interface {
+		ErrorTargets() []factoryapi.ErrorTarget
+	}
+	if !errors.As(err, &targetedErr) {
+		t.Fatalf("validation error %v did not expose structured targets", err)
+	}
+
+	targets := targetedErr.ErrorTargets()
+	if len(targets) != 1 {
+		t.Fatalf("validation error targets = %#v, want one target", targets)
+	}
+	target := targets[0]
+	if target.Kind != factorySessionValidationTargetKind {
+		t.Fatalf("validation target kind = %q, want %q", target.Kind, factorySessionValidationTargetKind)
+	}
+	if target.Id == nil || *target.Id != wantReason {
+		t.Fatalf("validation target id = %#v, want %q", target.Id, wantReason)
+	}
+	if target.Field == nil || *target.Field != wantField {
+		t.Fatalf("validation target field = %#v, want %q", target.Field, wantField)
 	}
 }
